requirements.txt
qrcode[pil]==8.0
Pillow